+++ vscode/src/completions/tree-sitter/queries/test-data/blocks.snap.ts
@@ -8,7 +8,7 @@
 // (_ ("{")) @blocks
 // 
 // [(try_statement)
-//  (if_statement)] @parents
+// (if_statement)] @parents
 // ------------------------------------
 
   interface kek {