+++ vscode/src/services/LocalAppFsPaths.ts
@@ -1,3 +1,8 @@
+export const LOCAL_APP_SETTINGS_DIR = new Map([
+    ['darwin', '~/Library/Application Support/com.sourcegraph.cody/'],
+    ['linux', '~/.local/share/com.sourcegraph.cody/'],
+])
+
 export const LOCAL_APP_LOCATIONS: LocalAppPaths = {
     darwin: [
         {